Have the large screen layout worked out where there are five sections:    Header    Left Sidebar    Right Sidebar    Main or Content    Footer
The Secondary Menu and Contact List are displayed in the Left Sidebar.
The Right Sidebar is currently empty and is used for spacing the
Main section.

Removed the menu item "Makerspace" from the main menu and placed it in
a secondary menu named "about-menu".  Added a Location page and a place
holder for a Contact Us page (menu item with no link).

Still need to add the search icon back in as well as doing some changes
to the menu system.  This is enough for a working demo that does not
yet support phone screens.