Added code so that the mobile submenus appear on the left side of the parent menu. This cannot be done in HTML or CSS and requires JavaScript.  To place a child element on the left of its parent requires knowing the width of the child.  CSS only provides access to the width of the parent container.  JavaScript can access the child width and change the style values (in this case 'left') to move the child (submenu).
This checkin is not the final checkin for this feature.  The JavaScript still contains
debugging code that is commented out and needs comments and refinement.  All the
JavaScript used by the menuus currently resides in the masthead.html.  At some point,
the JavaScript should be moved into its own file(s).

Future Feature: Use JavaScript to determine whether the submenu should be placed on
the right (default) or the left of the parent menu based on available space.